Check implementation against Metadata trait; track jump_targets.
Now that etk-asm has metadata about how many stack elements each Op
consumes/produces, and how each op can exit a block
(jumps/returns/etc.), we can check that against the simulation in
etk-analyze. Probably has a runtime cost, but I'd rather know if an Op
isn't implemented correctly.

Also, because we need it in the next step of generating a control flow
graph, each annotated block tracks whether it starts on a jumpdest or
not.